rework middleware
