+++ src/api-mongodb/pytest.ini
@@ -1,51 +1,6 @@
 [pytest]
 addopts = --maxfail=1
 testpaths =
-    test/api/test_main.py
+    test/api/test_hello.py
     test/api/test_auth.py
     test/api/internal/mongodb/entity/test_user.py
-    test/api/internal/mongodb/entity/test_space.py
-    test/api/internal/mongodb/entity/test_role.py
-    test/api/internal/mongodb/entity/test_document.py
-    test/api/internal/mongodb/entity/test_page.py
-	test/api/internal/mongodb/entity/test_widget.py
-    test/api/internal/mongodb/entity/test_directory.py
-    test/api/internal/mongodb/entity/test_index.py
-    test/api/internal/mongodb/entity/test_pipeline.py
-    test/api/internal/mongodb/entity/test_datasource.py
-    test/api/internal/mongodb/entity/test_api.py
-    test/api/internal/mongodb/entity/test_stage.py
-
-    test/api/internal/mongodb/relation/test_roles_users.py
-    test/api/internal/mongodb/relation/test_role_document.py
-    test/api/internal/mongodb/relation/test_role_space.py
-    test/api/internal/mongodb/relation/test_role_page.py
-    test/api/internal/mongodb/relation/test_role_widget.py
-    test/api/internal/mongodb/relation/test_role_directory.py
-    test/api/internal/mongodb/relation/test_role_index.py
-    test/api/internal/mongodb/relation/test_role_pipeline.py
-    test/api/internal/mongodb/relation/test_role_datasource.py
-    test/api/internal/mongodb/relation/test_role_api.py
-    test/api/internal/mongodb/relation/test_role_stage.py
-
-    test/api/internal/mongodb/relation/test_space_page.py
-    test/api/internal/mongodb/relation/test_space_directory.py
-    test/api/internal/mongodb/relation/test_space_index.py
-    test/api/internal/mongodb/relation/test_space_pipeline.py
-    test/api/internal/mongodb/relation/test_space_datasource.py
-    test/api/internal/mongodb/relation/test_space_api.py
-
-    test/api/internal/mongodb/relation/test_directory_directory.py
-    test/api/internal/mongodb/relation/test_directory_document.py
-
-    test/api/internal/mongodb/relation/test_pipeline_document.py
-    test/api/internal/mongodb/relation/test_pipeline_stage.py
-    test/api/internal/mongodb/relation/test_pipeline_index.py
-
-    test/api/internal/mongodb/relation/test_stage_api.py
-
-    test/api/internal/mongodb/relation/test_page_widget.py
-
-    test/api/internal/mongodb/relation/test_widget_datasource.py
-
-    test/api/internal/mongodb/relation/test_datasource_index.py
